+++ pkg/types/entities/eth_transaction.go
@@ -14,6 +14,15 @@ const (
 	DynamicFeeTxType TransactionType = "dynamic_fee"
 )
 
+type PrivacyFlag int
+
+const (
+	PrivacyFlagSP  PrivacyFlag = iota
+	PrivacyFlagPP  PrivacyFlag = 1
+	PrivacyFlagMPP PrivacyFlag = 2
+	PrivacyFlagPSV PrivacyFlag = 3
+)
+
 type ETHTransaction struct {
 	Hash            string           `json:"hash,omitempty" validate:"omitempty,isHex" example:"0xd41551c714c8ec769d2edad9adc250ae955d263da161bf59142b7500eea6715e"`
 	From            string           `json:"from,omitempty" validate:"omitempty,eth_addr" example:"0x1abae27a0cbfb02945720425d3b80c7e09728534"`
@@ -29,8 +38,10 @@ type ETHTransaction struct {
 	Data            string           `json:"data,omitempty" validate:"omitempty,isHex" example:"0xfe378324abcde723"`
 	Raw             string           `json:"raw,omitempty" validate:"omitempty,isHex" example:"0xfe378324abcde723"`
 	PrivateFrom     string           `json:"privateFrom,omitempty" validate:"omitempty,base64" example:"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo="`
-	PrivateFor      []string         `json:"privateFor,omitempty" validate:"omitempty,dive,base64" example:"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=,B1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo="`
-	PrivacyGroupID  string           `json:"privacyGroupID,omitempty" example:"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo="`
+	PrivateFor      []string         `json:"privateFor,omitempty" validate:"omitempty,min=1,unique,dive,base64" example:"[A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=,B1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=]"`
+	MandatoryFor    []string         `json:"mandatoryFor,omitempty" validate:"omitempty,min=1,unique,dive,base64" example:"[A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=,B1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=]"`
+	PrivacyGroupID  string           `json:"privacyGroupId,omitempty" validate:"omitempty,base64" example:"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo="`
+	PrivacyFlag     PrivacyFlag      `json:"privacyFlag,omitempty" validate:"omitempty,isPrivacyFlag" example:"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo="`
 	EnclaveKey      string           `json:"enclaveKey,omitempty" example:"0xd41551c714c8ec769d2edad9adc250ae955d263da161bf59142b7500eea6715eadc250ae955d263da161bf59142b7500eea6715e"`
 	CreatedAt       time.Time        `json:"createdAt,omitempty" example:"2020-07-09T12:35:42.115395Z"`
 	UpdatedAt       time.Time        `json:"updatedAt,omitempty" example:"2020-07-09T12:35:42.115395Z"`